Private types

Private types in module PassiveTracerFlows:

Private types in module PassiveTracerFlows:

PassiveTracerFlows.TracerAdvectionDiffusion.AbstractAdvectingFlow — Type

Abstract super type for an advecting flow.

PassiveTracerFlows.TracerAdvectionDiffusion.ConstDiffSteadyFlowParams1D — Type

struct ConstDiffSteadyFlowParams1D{T} <: AbstractSteadyFlowParams

A struct containing the parameters for a constant diffusivity problem with steady flow in one dimensions. Included are:

κ::Any

isotropic horizontal diffusivity coefficient

κh::Any

isotropic hyperdiffusivity coefficient

nκh::Int64

isotropic hyperdiffusivity order

u::Any

x-component of advecting flow

PassiveTracerFlows.TracerAdvectionDiffusion.ConstDiffSteadyFlowParams2D — Type

struct ConstDiffSteadyFlowParams2D{T} <: AbstractSteadyFlowParams

A struct containing the parameters for a constant diffusivity problem with steady flow in two dimensions. Included are:

η::Any

isotropic horizontal diffusivity coefficient

κ::Any

isotropic vertical diffusivity coefficient

κh::Any

isotropic hyperdiffusivity coefficient

nκh::Int64

isotropic hyperdiffusivity order

u::Any

x-component of advecting flow

v::Any

y-component of advecting flow

PassiveTracerFlows.TracerAdvectionDiffusion.ConstDiffTimeVaryingFlowParams1D — Type

struct ConstDiffTimeVaryingFlowParams1D{T} <: AbstractTimeVaryingFlowParams

A struct containing the parameters for a constant diffusivity problem with time-varying flow in one dimension. Included are:

κ::Any

isotropic horizontal diffusivity coefficient

κh::Any

isotropic hyperdiffusivity coefficient

nκh::Int64

isotropic hyperdiffusivity order

u::Function

function returning the x-component of advecting flow

PassiveTracerFlows.TracerAdvectionDiffusion.ConstDiffTimeVaryingFlowParams2D — Type

struct ConstDiffTimeVaryingFlowParams2D{T} <: AbstractTimeVaryingFlowParams

A struct containing the parameters for a constant diffusivity problem with time-varying flow in two dimensions. Included are:

η::Any

isotropic horizontal diffusivity coefficient

κ::Any

isotropic vertical diffusivity coefficient

κh::Any

isotropic hyperdiffusivity coefficient

nκh::Int64

isotropic hyperdiffusivity order

u::Function

function returning the x-component of advecting flow

v::Function

function returning the y-component of advecting flow

PassiveTracerFlows.TracerAdvectionDiffusion.ConstDiffTurbulentFlowParams — Type

struct ConstDiffTurbulentFlowParams{T} <: AbstractTurbulentFlowParams

A struct containing the parameters for a constant diffusivity problem with flow obtained from a GeophysicalFlows.MultiLayerQG problem.

η::Any

isotropic horizontal diffusivity coefficient

κ::Any

isotropic vertical diffusivity coefficient

κh::Any

isotropic hyperdiffusivity coefficient

nκh::Int64

isotropic hyperdiffusivity order

nlayers::Int64

number of layers in which the tracer is advected-diffused

tracer_release_time::Any

flow time prior to releasing tracer

MQGprob::FourierFlows.Problem

MultiLayerQG.Problem to generate the advecting flow

PassiveTracerFlows.TracerAdvectionDiffusion.ConstDiffTurbulentFlowParams — Method

ConstDiffTurbulentFlowParams(η, κ, tracer_release_time, MQGprob)

Return the parameters params for a constant diffusivity problem with flow obtained from a GeophysicalFlows.MultiLayerQG problem.

PassiveTracerFlows.TracerAdvectionDiffusion.Vars1D — Type

struct Vars1D{Aphys, Atrans} <: AbstractVars

The variables for a 1D TracerAdvectionDiffussion problem.

c

tracer concentration

cx

tracer concentration $x$-derivative, $∂c/∂x$

ch

Fourier transform of tracer concentration

cxh

Fourier transform of tracer concentration $x$-derivative, $∂c/∂x$

PassiveTracerFlows.TracerAdvectionDiffusion.Vars2D — Type

struct Vars2D{Aphys, Atrans} <: AbstractVars

The variables for a 2D TracerAdvectionDiffussion problem.

c

tracer concentration

cx

tracer concentration $x$-derivative, $∂c/∂x$

cy

tracer concentration $y$-derivative, $∂c/∂y$

ch

Fourier transform of tracer concentration

cxh

Fourier transform of tracer concentration $x$-derivative, $∂c/∂x$

cyh

Fourier transform of tracer concentration $y$-derivative, $∂c/∂y$