Private types

Private types in module PassiveTracerFlows:

Private types in module PassiveTracerFlows:

PassiveTracerFlows.TracerAdvectionDiffusion.ConstDiffParams — Type

struct ConstDiffParams{T} <: AbstractConstDiffParams

A struct containing the parameters for a constant diffusivity problem with time-varying flow. Included are:

η::Any

isotropic horizontal diffusivity coefficient

κ::Any

isotropic vertical diffusivity coefficient

κh::Any

isotropic hyperdiffusivity coefficient

nκh::Int64

isotropic hyperdiffusivity order

u::Function

function returning the x-component of advecting flow

v::Function

function returning the y-component of advecting flow

PassiveTracerFlows.TracerAdvectionDiffusion.ConstDiffParams — Method

ConstDiffParams(η, κ, u, v)

The constructor for the params struct for constant diffusivity problem and time-varying flow.

PassiveTracerFlows.TracerAdvectionDiffusion.ConstDiffSteadyFlowParams — Type

struct ConstDiffParams{T} <: AbstractConstDiffParams

A struct containing the parameters for a constant diffusivity problem with steady flow. Included are:

η::Any

isotropic horizontal diffusivity coefficient

κ::Any

isotropic vertical diffusivity coefficient

κh::Any

isotropic hyperdiffusivity coefficient

nκh::Int64

isotropic hyperdiffusivity order

u::Any

x-component of advecting flow

v::Any

y-component of advecting flow

PassiveTracerFlows.TracerAdvectionDiffusion.ConstDiffSteadyFlowParams — Method

ConstDiffSteadyFlowParams(η, κ, κh, nκh, u::Function, v::Function, grid)
ConstDiffSteadyFlowParams(η, κ, u, v, grid)

The constructor for the params struct for constant diffusivity problem and steady flow.

PassiveTracerFlows.TracerAdvectionDiffusion.TurbulentFlowParams — Type

struct TurbulentFlowParams{T} <: AbstractTurbulentFlowParams

A struct containing the parameters for a constant diffusivity problem with flow obtained from a GeophysicalFlows.MultiLayerQG problem.

η::Any

isotropic horizontal diffusivity coefficient

κ::Any

isotropic vertical diffusivity coefficient

κh::Any

isotropic hyperdiffusivity coefficient

nκh::Int64

isotropic hyperdiffusivity order

nlayers::Int64

number of layers in which the tracer is advected-diffused

tracer_release_time::Any

flow time prior to releasing tracer

MQGprob::FourierFlows.Problem

MultiLayerQG.Problem to generate the advecting flow

PassiveTracerFlows.TracerAdvectionDiffusion.TurbulentFlowParams — Method

TurbulentFlowParams(η, κ, tracer_release_time, MQGprob)

The constructor for the params for a constant diffusivity problem with flow obtained from a GeophysicalFlows.MultiLayerQG problem.

PassiveTracerFlows.TracerAdvectionDiffusion.Vars — Type

struct Vars{Aphys, Atrans} <: AbstractVars

The variables for TracerAdvectionDiffussion problems.

c

tracer concentration

cx

tracer concentration x-derivative, ∂c/∂x

cy

tracer concentration y-derivative, ∂c/∂y

ch

Fourier transform of tracer concentration

cxh

Fourier transform of tracer concentration x-derivative, ∂c/∂x

cyh

Fourier transform of tracer concentration y-derivative, ∂c/∂y

PassiveTracerFlows.TracerAdvectionDiffusion.Vars — Method

Vars(dev, grid)

Return the variables vars for a constant diffusivity problem on grid and device dev.