Private types

Private types in module PassiveTracerFlows:

Private types in module PassiveTracerFlows:

PassiveTracerFlows.TracerAdvectionDiffusion.ConstDiffParams — Type

ConstDiffParams(η, κ, κh, nκh, u, v)
ConstDiffParams(η, κ, u, v)

Returns the params for constant diffusivity problem with time-varying flow.

PassiveTracerFlows.TracerAdvectionDiffusion.ConstDiffSteadyFlowParams — Type

ConstDiffSteadyFlowParams(η, κ, κh, nκh, u, v, g)
ConstDiffSteadyFlowParams(η, κ, u, v, g)

Returns the params for constant diffusivity problem with time-steady flow.

PassiveTracerFlows.TracerAdvectionDiffusion.Vars — Method

Vars(g)

Returns the vars for constant diffusivity problem on grid g.