Private types

Private types in module PassiveTracerFlows:

Private types in module TracerAdvectionDiffusion:

PassiveTracerFlows.TracerAdvectionDiffusion.AbstractAdvectingFlow — Type

Abstract super type for an advecting flow.

PassiveTracerFlows.TracerAdvectionDiffusion.ConstDiffSteadyFlowParams1D — Type

struct ConstDiffSteadyFlowParams1D{T} <: AbstractSteadyFlowParams

The parameters of a constant diffusivity problem with steady flow in one dimension.

κ::Any

$x$-diffusivity coefficient

κh::Any

isotropic hyperdiffusivity coefficient

nκh::Int64

isotropic hyperdiffusivity order

u::Any

$x$-component of advecting flow

PassiveTracerFlows.TracerAdvectionDiffusion.ConstDiffSteadyFlowParams2D — Type

struct ConstDiffSteadyFlowParams2D{T} <: AbstractSteadyFlowParams

The parameters for a constant diffusivity problem with steady flow in two dimensions.

κ::Any

$x$-diffusivity coefficient

η::Any

$y$-diffusivity coefficient

κh::Any

isotropic hyperdiffusivity coefficient

nκh::Int64

isotropic hyperdiffusivity order

u::Any

$x$-component of advecting flow

v::Any

$y$-component of advecting flow

PassiveTracerFlows.TracerAdvectionDiffusion.ConstDiffSteadyFlowParams3D — Type

struct ConstDiffSteadyFlowParams3D{T} <: AbstractSteadyFlowParams

The parameters for a constant diffusivity problem with steady flow in three dimensions.

κ::Any

$x$-diffusivity coefficient

η::Any

$y$-diffusivity coefficient

ι::Any

$z$-diffusivity coefficient

κh::Any

isotropic hyperdiffusivity coefficient

nκh::Int64

isotropic hyperdiffusivity order

u::Any

$x$-component of advecting flow

v::Any

$y$-component of advecting flow

w::Any

$z$-component of advecting flow

PassiveTracerFlows.TracerAdvectionDiffusion.ConstDiffTimeVaryingFlowParams1D — Type

struct ConstDiffTimeVaryingFlowParams1D{T} <: AbstractTimeVaryingFlowParams

The parameters of a constant diffusivity problem with time-varying flow in one dimension.

κ::Any

diffusivity coefficient

κh::Any

hyperdiffusivity coefficient

nκh::Int64

hyperdiffusivity order

u::Function

function returning the $x$-component of advecting flow

PassiveTracerFlows.TracerAdvectionDiffusion.ConstDiffTimeVaryingFlowParams2D — Type

struct ConstDiffTimeVaryingFlowParams2D{T} <: AbstractTimeVaryingFlowParams

The parameters of a constant diffusivity problem with time-varying flow in two dimensions.

κ::Any

$x$-diffusivity coefficient

η::Any

$y$-diffusivity coefficient

κh::Any

isotropic hyperdiffusivity coefficient

nκh::Int64

isotropic hyperdiffusivity order

u::Function

function returning the $x$-component of advecting flow

v::Function

function returning the $y$-component of advecting flow

PassiveTracerFlows.TracerAdvectionDiffusion.ConstDiffTimeVaryingFlowParams3D — Type

struct ConstDiffTimeVaryingFlowParams3D{T} <: AbstractTimeVaryingFlowParams

The parameters of a constant diffusivity problem with time-varying flow in three dimensions.

κ::Any

$x$-diffusivity coefficient

η::Any

$y$-diffusivity coefficient

ι::Any

$z$-diffusivity coefficient

κh::Any

isotropic hyperdiffusivity coefficient

nκh::Int64

isotropic hyperdiffusivity order

u::Function

function returning the $x$-component of advecting flow

v::Function

function returning the $y$-component of advecting flow

w::Function

function returning the $z$-component of advecting flow

PassiveTracerFlows.TracerAdvectionDiffusion.ConstDiffTurbulentFlowParams — Type

struct ConstDiffTurbulentFlowParams{T} <: AbstractTurbulentFlowParams

The parameters of a constant diffusivity problem with flow obtained from a GeophysicalFlows.MultiLayerQG problem.

κ::Any

$x$-diffusivity coefficient

η::Any

$y$-diffusivity coefficient

κh::Any

isotropic hyperdiffusivity coefficient

nκh::Int64

isotropic hyperdiffusivity order

nlayers::Int64

number of layers in which the tracer is advected-diffused

tracer_release_time::Any

flow time prior to releasing tracer

MQGprob::FourierFlows.Problem

MultiLayerQG.Problem to generate the advecting flow

PassiveTracerFlows.TracerAdvectionDiffusion.ConstDiffTurbulentFlowParams — Method

ConstDiffTurbulentFlowParams(κ, η, tracer_release_time, MQGprob)

Return the parameters params for a constant diffusivity problem with flow obtained from a GeophysicalFlows.MultiLayerQG problem.

PassiveTracerFlows.TracerAdvectionDiffusion.Vars1D — Type

struct Vars1D{Aphys, Atrans} <: AbstractVars

The variables of a 1D TracerAdvectionDiffussion problem.

c

tracer concentration

cx

tracer concentration $x$-derivative, $∂c/∂x$

ch

Fourier transform of tracer concentration

cxh

Fourier transform of tracer concentration $x$-derivative, $∂c/∂x$

PassiveTracerFlows.TracerAdvectionDiffusion.Vars2D — Type

struct Vars2D{Aphys, Atrans} <: AbstractVars

The variables of a 2D TracerAdvectionDiffussion problem.

c

tracer concentration

cx

tracer concentration $x$-derivative, $∂c/∂x$

cy

tracer concentration $y$-derivative, $∂c/∂y$

ch

Fourier transform of tracer concentration

cxh

Fourier transform of tracer concentration $x$-derivative, $∂c/∂x$

cyh

Fourier transform of tracer concentration $y$-derivative, $∂c/∂y$

PassiveTracerFlows.TracerAdvectionDiffusion.Vars3D — Type

struct Vars3D{Aphys, Atrans} <: AbstractVars

The variables of a 3D TracerAdvectionDiffussion problem.

c

tracer concentration

cx

tracer concentration $x$-derivative, $∂c/∂x$

cy

tracer concentration $y$-derivative, $∂c/∂y$

cz

tracer concentration $z$-derivative, $∂c/∂z$

ch

Fourier transform of tracer concentration

cxh

Fourier transform of tracer concentration $x$-derivative, $∂c/∂x$

cyh

Fourier transform of tracer concentration $y$-derivative, $∂c/∂y$

czh

Fourier transform of tracer concentration $z$-derivative, $∂c/∂z$